added VHDL for text_datapath (and testbench) and text_package
Update: Most recent git-push from me is a component (and relevant package) meant to feed the correct information to char_rom based on the position of the pixel currently refreshed on screen (without scaling).

This component currently SHOULD work together with char_rom and VGA_sync. However, ideally, it should be a part of GUI and eventually Graphics.

I am 40% sure it will work in its current state. Not yet synthesized. Further testing required in Quartus and on DE0 board.